requirements.txt
#  pip install -r requirements.txt
#  sudo apt-get install python3-pip
#  pip3 install flask_cors
#  pip3 install flask
flask == 1.1.1
flask-cors == 3.0.8
Jinja2 == 2.10.1
click == 7.0
itsdangerous == 1.1.0
Werkzeug == 0.16.0
MarkupSafe == 1.1.1
reverse_geocoder
geocoder == 1.38.1
PyJWT == 1.7.1
pymongo == 3.9.0
dnspython == 1.16.0
google-cloud-storage == 1.20.0
